refactor(plugs): renamed entity folder to schemas
