+++ api/services/plugs/src/services/services/services.service.ts
@@ -1,6 +1,6 @@
 import { Injectable } from '@nestjs/common';
 import { InjectModel } from '@nestjs/mongoose';
-import { Service, ServiceDocument } from '../entity/service.schema';
+import { Service, ServiceDocument } from '../schemas/service.schema';
 import { Model } from 'mongoose';
 import {
   ActionDescription,
